Add option for removeing the leading path element.
Add :strip-leading-path to the :tar options hashmap.
If :strip-leading-path is set to true, then the leading directory path
element (in this plugin's case the project name) will be stripped.

This gives the same result as running the folowing in the command line:
tar czf lein-tar-3.2.0.tgz -C lein-tar-3.2.0/ .

Example:

:strip-leading-path         Tar archive
-------------------------------------------------------------------
false                      lein-tar-3.2.0/
                           lein-tar-3.2.0/bin
                           lein-tar-3.2.0/bin/heyo
                           lein-tar-3.2.0/etc
                           lein-tar-3.2.0/etc/foo.conf
                           lein-tar-3.2.0/lib/
                           lein-tar-3.2.0/lib/lein-tar-3.2.0.jar

true                       bin
                           bin/heyo
                           etc
                           etc/foo.conf
                           lib/
                           lib/lein-tar-3.2.0.jar